requirements.txt
PySide6>=6.0.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.20.0
